Move metastore specifics under their own object
